Include all Teleport CAs in the rotation guide
Closes #37125

This change clarifies the differences between Teleport CAs in the CA
Rotation guide. Since the Database CA Rotation guide is almost identical
to the original CA Rotation guide, this change merges the two guides.

To include relevant information for each CA, this change adds a step to
help the user choose a CA to rotate, and includes specific information for
each CA in that step.

Another possibility would be to spread information for each CA across
the instructions for each transition phase. However, since each CA
requires a different set of checks/updates in different phases, it makes
sense to make the instructions for each CA more free-flowing while
restricting them to a single, early section.

Most information about checking the status of a rotation in the original
guide had to do with the `host` CA, and this change moves that
information into an H3 related to that CA.